:truck: refactor: add index and modules for each domain (#13)
