--- src/domain/album/album.controller.ts
@@ -1,41 +1,0 @@
-import {
-  Controller,
-  Get,
-  Post,
-  Body,
-  Patch,
-  Param,
-  Delete,
-} from '@nestjs/common';
-import { AlbumService } from '../family/service';
-import { CreateAlbumDto, UpdateAlbumDto } from '../family/dto';
-
-@Controller('album')
-export class AlbumController {
-  constructor(private readonly albumService: AlbumService) {}
-
-  @Post()
-  create(@Body() createAlbumDto: CreateAlbumDto) {
-    return this.albumService.create(createAlbumDto);
-  }
-
-  @Get()
-  findAll() {
-    return this.albumService.findAll();
-  }
-
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.albumService.findOne(+id);
-  }
-
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateAlbumDto: UpdateAlbumDto) {
-    return this.albumService.update(+id, updateAlbumDto);
-  }
-
-  @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.albumService.remove(+id);
-  }
-}
