--- src/domain/album/album.service.ts
@@ -1,25 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { CreateAlbumDto, UpdateAlbumDto } from '../family/dto';
-
-@Injectable()
-export class AlbumService {
-  create(createAlbumDto: CreateAlbumDto) {
-    return 'This action adds a new album';
-  }
-
-  findAll() {
-    return `This action returns all album`;
-  }
-
-  findOne(id: number) {
-    return `This action returns a #${id} album`;
-  }
-
-  update(id: number, updateAlbumDto: UpdateAlbumDto) {
-    return `This action updates a #${id} album`;
-  }
-
-  remove(id: number) {
-    return `This action removes a #${id} album`;
-  }
-}
