--- src/domain/album/dto/create-album.dto.ts
@@ -1,1 +1,0 @@
-export class CreateAlbumDto {}
