--- src/domain/album/dto/update-album.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateAlbumDto } from './create-album.dto';
-
-export class UpdateAlbumDto extends PartialType(CreateAlbumDto) {}
