+++ src/domain/family-member/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './request/create-family-member.dto';
+export * from './request/update-family-member.dto';
+export * from './response/response-family-member.dto';
