+++ src/domain/family-member/index.ts
@@ -1,3 +1,5 @@
 export * from './dto/request/create-family-member.dto';
 export * from './dto/request/update-family-member.dto';
 export * from './dto/response/response-family-member.dto';
+export * from './family-member.service';
+export * from './family-member.controller';
