+++ src/domain/family-schedule/family-schedule.controller.ts
@@ -7,8 +7,8 @@ import {
   Param,
   Delete,
 } from '@nestjs/common';
-import { FamilyScheduleService } from '../family/service';
-import { CreateFamilyScheduleDto, UpdateFamilyScheduleDto } from '../family/dto';
+import { FamilyScheduleService } from '../family-schedule';
+import { CreateFamilyScheduleDto, UpdateFamilyScheduleDto } from './dto';
 
 @Controller('family-schedule')
 export class FamilyScheduleController {
