+++ src/domain/family-schedule/index.ts
@@ -1,0 +1,6 @@
+export * from './family-schedule.module';
+export * from './family-schedule.service';
+export * from './family-schedule.controller';
+export * from './dto/request/create-family-schedule.dto';
+export * from './dto/request/update-family-schedule.dto';
+export * from './dto/response/response-family-schedule.dto';
