+++ src/domain/family/family.module.ts
@@ -1,21 +1,13 @@
 import { Module } from '@nestjs/common';
-import {
-  FamilyService,
-
-} from './family.service';
-import {
-  FamilyController,
-
-} from './family.controller';
+import { FamilyService } from './family.service';
+import { FamilyController } from './family.controller';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Family } from '../../infra/entities';
 
 @Module({
-  controllers: [
-    FamilyController,
+  imports: [TypeOrmModule.forFeature([Family])],
 
-  ],
-  providers: [
-    FamilyService,
-
-  ],
+  controllers: [FamilyController],
+  providers: [FamilyService],
 })
 export class FamilyModule {}
