+++ src/domain/family/index.ts
@@ -1,0 +1,6 @@
+export * from './dto/request/create-family.dto';
+export * from './dto/request/update-family.dto';
+export * from './dto/response/response-family.dto';
+export * from './family.service';
+export * from './family.controller';
+export * from './family.module';
