+++ src/domain/photo/dto/index.ts
