+++ src/domain/photo/index.ts
@@ -1,0 +1,6 @@
+export * from './dto/request/create-photo.dto';
+export * from './dto/request/update-photo.dto';
+export * from './dto/response/response-photo.dto';
+export * from './photo.module';
+export * from './photo.service';
+export * from './photo.controller';
