+++ src/domain/post/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './request/create-post.dto';
+export * from './request/update-post.dto';
+export * from './response/response-post.dto';
