+++ src/domain/post/dto/response/response-post.dto.ts
@@ -1,0 +1,1 @@
+export class ResponsePostDto {}
