+++ src/domain/post/post.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { PostController } from './post.controller';
+import { PostService } from './post.service';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Post } from '../../infra/entities';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Post])],
+  controllers: [PostController],
+  providers: [PostService],
+})
+export class PostModule {}
