+++ src/domain/user/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './request/create-user.dto';
+export * from './request/update-user.dto';
