+++ src/domain/user/user.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from '../../infra/entities/user.entity';
+import { User } from '../../infra/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
